Corrected type assignment and subtype detection
Inspections were previously attempting to determine if a type was
assignable to another type simply by checking if the former is a
subtype of latter. This was only correct in the simple case, and
to work around this to a degree several types were indicating they
were a subtype of another, even when they weren't.

To correct this there's now a new assignableFrom() method on each
Ty. Simple types fallback to a subtype check, but more complex
types (generics, generic parameters and unions) implement this
method to more accurately enforce type correctness.

Subype detection has now been corrected for generics and unions
i.e. All members of a union must be a subtype of a type for the
union to be considered a subtype e.g. string|number is *not* a
subtype of string. Additionally, generics are *not* a subtype of
another generic unless their parameters match exactly.

Additionally, TyParameter is now resolved where appropriate, rather
than simply resolving as Ty, thus type assignment inspections now
work for generics.