Major refactoring to allow unsuccesful type resolution to bubble.
The previous infrastructure was extremely problematic as in the
cast of type resolution/inference failing (for whatever reason)
Ty.UNKNOWN was returned. This is a major concern because it was
impossible to discern between a resolution failure and a type that
was explictly typed as Ty.UNKNOWN.

Now we're able to handle unsuccessful type resolution on a case
by case basis.

This also paves the way for (at some point) distinguishing between
'unknown' and 'any'.