Formalised type widening behaviour for overidden fields/methods.
Essentially, an explicit type annotation will always take
precedence over an implicitly inferred type. Thus, if you wish
a sub-class to narrow an overridden type, you ought to do so
explicitly with a type annotation.

The only time the child's type will automatically narrow a
parent's type is if the parent did not provide an explicit type
annotation for the field/method.