Fixed type of anonymous table members derived from generic params.
Previously the type worked within the scope of the defining
function, but once returned the type failed to resolve and was
essentially unusable. Explicit types (aliases or shapes) were a
workaround, but this commit resolves the problem.

Additionally, we've improved the displayed type for anonymous
tables. Rather than just seeing "table" you know see the members
and their types in both error messages and when hovering over
variables.